Change sig

+++ Iced/Intel/Encoder.cs
@@ -950,21 +950,21 @@ namespace Iced.Intel {
 		void WriteOpCode() {
 			var opCode = OpCode;
 			if (opCode <= 0x000000FF)
-				WriteByte((byte)opCode);
+				WriteByte(opCode);
 			else if (opCode <= 0x0000FFFF) {
-				WriteByte((byte)(opCode >> 8));
-				WriteByte((byte)opCode);
+				WriteByte(opCode >> 8);
+				WriteByte(opCode);
 			}
 			else if (opCode <= 0x00FFFFFF) {
-				WriteByte((byte)(opCode >> 16));
-				WriteByte((byte)(opCode >> 8));
-				WriteByte((byte)opCode);
+				WriteByte(opCode >> 16);
+				WriteByte(opCode >> 8);
+				WriteByte(opCode);
 			}
 			else {
-				WriteByte((byte)(opCode >> 24));
-				WriteByte((byte)(opCode >> 16));
-				WriteByte((byte)(opCode >> 8));
-				WriteByte((byte)opCode);
+				WriteByte(opCode >> 24);
+				WriteByte(opCode >> 16);
+				WriteByte(opCode >> 8);
+				WriteByte(opCode);
 			}
 		}
 
@@ -983,43 +983,43 @@ namespace Iced.Intel {
 				break;
 
 			case DisplSize.Size1:
-				WriteByte((byte)Displ);
+				WriteByte(Displ);
 				break;
 
 			case DisplSize.Size2:
 				diff4 = Displ;
-				WriteByte((byte)diff4);
-				WriteByte((byte)(diff4 >> 8));
+				WriteByte(diff4);
+				WriteByte(diff4 >> 8);
 				break;
 
 			case DisplSize.Size4:
 				diff4 = Displ;
-				WriteByte((byte)diff4);
-				WriteByte((byte)(diff4 >> 8));
-				WriteByte((byte)(diff4 >> 16));
-				WriteByte((byte)(diff4 >> 24));
+				WriteByte(diff4);
+				WriteByte(diff4 >> 8);
+				WriteByte(diff4 >> 16);
+				WriteByte(diff4 >> 24);
 				break;
 
 			case DisplSize.Size8:
 				diff4 = Displ;
-				WriteByte((byte)diff4);
-				WriteByte((byte)(diff4 >> 8));
-				WriteByte((byte)(diff4 >> 16));
-				WriteByte((byte)(diff4 >> 24));
+				WriteByte(diff4);
+				WriteByte(diff4 >> 8);
+				WriteByte(diff4 >> 16);
+				WriteByte(diff4 >> 24);
 				diff4 = DisplHi;
-				WriteByte((byte)diff4);
-				WriteByte((byte)(diff4 >> 8));
-				WriteByte((byte)(diff4 >> 16));
-				WriteByte((byte)(diff4 >> 24));
+				WriteByte(diff4);
+				WriteByte(diff4 >> 8);
+				WriteByte(diff4 >> 16);
+				WriteByte(diff4 >> 24);
 				break;
 
 			case DisplSize.RipRelSize4_Target32:
 				uint eip = (uint)currentRip + 4 + immSizes[(int)ImmSize];
 				diff4 = Displ - eip;
-				WriteByte((byte)diff4);
-				WriteByte((byte)(diff4 >> 8));
-				WriteByte((byte)(diff4 >> 16));
-				WriteByte((byte)(diff4 >> 24));
+				WriteByte(diff4);
+				WriteByte(diff4 >> 8);
+				WriteByte(diff4 >> 16);
+				WriteByte(diff4 >> 24);
 				break;
 
 			case DisplSize.RipRelSize4_Target64:
@@ -1028,10 +1028,10 @@ namespace Iced.Intel {
 				if (diff8 < int.MinValue || diff8 > int.MaxValue)
 					ErrorMessage = $"RIP relative distance is too far away: nextIp: 0x{rip:X16} target: 0x{DisplHi:X8}{Displ:X8}, diff = {diff8}, diff must fit in an Int32";
 				diff4 = (uint)diff8;
-				WriteByte((byte)diff4);
-				WriteByte((byte)(diff4 >> 8));
-				WriteByte((byte)(diff4 >> 16));
-				WriteByte((byte)(diff4 >> 24));
+				WriteByte(diff4);
+				WriteByte(diff4 >> 8);
+				WriteByte(diff4 >> 16);
+				WriteByte(diff4 >> 24);
 				break;
 
 			default:
@@ -1055,66 +1055,66 @@ namespace Iced.Intel {
 			case ImmSize.Size1:
 			case ImmSize.SizeIbReg:
 			case ImmSize.Size1OpCode:
-				WriteByte((byte)Immediate);
+				WriteByte(Immediate);
 				break;
 
 			case ImmSize.Size2:
 				value = Immediate;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
+				WriteByte(value);
+				WriteByte(value >> 8);
 				break;
 
 			case ImmSize.Size4:
 				value = Immediate;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
-				WriteByte((byte)(value >> 16));
-				WriteByte((byte)(value >> 24));
+				WriteByte(value);
+				WriteByte(value >> 8);
+				WriteByte(value >> 16);
+				WriteByte(value >> 24);
 				break;
 
 			case ImmSize.Size8:
 				value = Immediate;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
-				WriteByte((byte)(value >> 16));
-				WriteByte((byte)(value >> 24));
+				WriteByte(value);
+				WriteByte(value >> 8);
+				WriteByte(value >> 16);
+				WriteByte(value >> 24);
 				value = ImmediateHi;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
-				WriteByte((byte)(value >> 16));
-				WriteByte((byte)(value >> 24));
+				WriteByte(value);
+				WriteByte(value >> 8);
+				WriteByte(value >> 16);
+				WriteByte(value >> 24);
 				break;
 
 			case ImmSize.Size2_1:
 				value = Immediate;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
-				WriteByte((byte)ImmediateHi);
+				WriteByte(value);
+				WriteByte(value >> 8);
+				WriteByte(ImmediateHi);
 				break;
 
 			case ImmSize.Size1_1:
-				WriteByte((byte)Immediate);
-				WriteByte((byte)ImmediateHi);
+				WriteByte(Immediate);
+				WriteByte(ImmediateHi);
 				break;
 
 			case ImmSize.Size2_2:
 				value = Immediate;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
+				WriteByte(value);
+				WriteByte(value >> 8);
 				value = ImmediateHi;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
+				WriteByte(value);
+				WriteByte(value >> 8);
 				break;
 
 			case ImmSize.Size4_2:
 				value = Immediate;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
-				WriteByte((byte)(value >> 16));
-				WriteByte((byte)(value >> 24));
+				WriteByte(value);
+				WriteByte(value >> 8);
+				WriteByte(value >> 16);
+				WriteByte(value >> 24);
 				value = ImmediateHi;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
+				WriteByte(value);
+				WriteByte(value >> 8);
 				break;
 
 			case ImmSize.RipRelSize1_Target16:
@@ -1122,7 +1122,7 @@ namespace Iced.Intel {
 				diff2 = (short)((short)Immediate - ip);
 				if (diff2 < sbyte.MinValue || diff2 > sbyte.MaxValue)
 					ErrorMessage = $"Branch distance is too far away: nextIp: 0x{ip:X4} target: 0x{(ushort)Immediate:X4}, diff = {diff2}, diff must fit in an Int8";
-				WriteByte((byte)diff2);
+				WriteByte((uint)diff2);
 				break;
 
 			case ImmSize.RipRelSize1_Target32:
@@ -1130,7 +1130,7 @@ namespace Iced.Intel {
 				diff4 = (int)Immediate - (int)eip;
 				if (diff4 < sbyte.MinValue || diff4 > sbyte.MaxValue)
 					ErrorMessage = $"Branch distance is too far away: nextIp: 0x{eip:X8} target: 0x{Immediate:X8}, diff = {diff4}, diff must fit in an Int8";
-				WriteByte((byte)diff4);
+				WriteByte((uint)diff4);
 				break;
 
 			case ImmSize.RipRelSize1_Target64:
@@ -1138,14 +1138,14 @@ namespace Iced.Intel {
 				diff8 = (long)(((ulong)ImmediateHi << 32) | (ulong)Immediate) - (long)rip;
 				if (diff8 < sbyte.MinValue || diff8 > sbyte.MaxValue)
 					ErrorMessage = $"Branch distance is too far away: nextIp: 0x{rip:X16} target: 0x{ImmediateHi:X8}{Immediate:X8}, diff = {diff8}, diff must fit in an Int8";
-				WriteByte((byte)diff8);
+				WriteByte((uint)diff8);
 				break;
 
 			case ImmSize.RipRelSize2_Target16:
 				ip = (ushort)((uint)currentRip + 2);
 				value = Immediate - ip;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
+				WriteByte(value);
+				WriteByte(value >> 8);
 				break;
 
 			case ImmSize.RipRelSize2_Target32:
@@ -1154,8 +1154,8 @@ namespace Iced.Intel {
 				if (diff4 < short.MinValue || diff4 > short.MaxValue)
 					ErrorMessage = $"Branch distance is too far away: nextIp: 0x{eip:X8} target: 0x{Immediate:X8}, diff = {diff4}, diff must fit in an Int16";
 				value = (uint)diff4;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
+				WriteByte(value);
+				WriteByte(value >> 8);
 				break;
 
 			case ImmSize.RipRelSize2_Target64:
@@ -1164,17 +1164,17 @@ namespace Iced.Intel {
 				if (diff8 < short.MinValue || diff8 > short.MaxValue)
 					ErrorMessage = $"Branch distance is too far away: nextIp: 0x{rip:X16} target: 0x{ImmediateHi:X8}{Immediate:X8}, diff = {diff8}, diff must fit in an Int16";
 				value = (uint)diff8;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
+				WriteByte(value);
+				WriteByte(value >> 8);
 				break;
 
 			case ImmSize.RipRelSize4_Target32:
 				eip = (uint)currentRip + 4;
 				value = Immediate - eip;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
-				WriteByte((byte)(value >> 16));
-				WriteByte((byte)(value >> 24));
+				WriteByte(value);
+				WriteByte(value >> 8);
+				WriteByte(value >> 16);
+				WriteByte(value >> 24);
 				break;
 
 			case ImmSize.RipRelSize4_Target64:
@@ -1183,10 +1183,10 @@ namespace Iced.Intel {
 				if (diff8 < int.MinValue || diff8 > int.MaxValue)
 					ErrorMessage = $"Branch distance is too far away: nextIp: 0x{rip:X16} target: 0x{ImmediateHi:X8}{Immediate:X8}, diff = {diff8}, diff must fit in an Int32";
 				value = (uint)diff8;
-				WriteByte((byte)value);
-				WriteByte((byte)(value >> 8));
-				WriteByte((byte)(value >> 16));
-				WriteByte((byte)(value >> 24));
+				WriteByte(value);
+				WriteByte(value >> 8);
+				WriteByte(value >> 16);
+				WriteByte(value >> 24);
 				break;
 
 			case ImmSize.Last:
@@ -1196,8 +1196,8 @@ namespace Iced.Intel {
 		}
 
 		[MethodImpl(MethodImplOptions.AggressiveInlining)]
-		internal void WriteByte(byte value) {
-			writer.WriteByte(value);
+		internal void WriteByte(uint value) {
+			writer.WriteByte((byte)value);
 			currentRip++;
 		}
